Search in a Rotated Array
Search in a Rotated Array
Given a sorted and rotated array A of N distinct elements which is rotated at some point, and given an element key. The task is to find the index of the given element key in the array A.

Expected Time Complexity O(log N).

Expected Auxiliary Space O(1).

Input Format
The first line contains two integers N and key.

The second line contains N spaced integers, elements of A.

Output Format
Print the index of the given key in the array. If the key is not present print -1.

Example 1
Input

9 10
5 6 7 8 9 10 1 2 3
Output

5
Explanation

10 is found at index 5

Example 2
Input

7 10
4 7 8 9 10 1 3
Output

4
Explanation

10 is found at index 4

Constraints
1 ≤ N ≤ 10^7

0 ≤ A[i] ≤ 10^8

1 ≤ key ≤ 10^8