Minimum Element in Sorted and Rotated Array
Minimum Element in Sorted and Rotated Array
Given an array A which is sorted in ascending order and contains N elements. Also, this array is rotated at some unknown point. The task is to find the minimum element in it.

Input
The first line consists of an integer N, where N is the size of array.
The second line contains N space separated integers denoting array elements.

Constraints:
1 <= N <= 10^5
-10^6 <= A[i] <= 10^5

Output
Print the minimum element in the array in new line.

Example
Input:
10
2 3 4 5 6 7 8 9 10 1
Output:
1 
Explanation:
The array is rotated once anti-clockwise. So minimum element is at last index (n-1) which is 1.
Input:
8
10 20 30 45 50 60 4 6
Output:
4 
Explanation:
The array is rotated and the minimum element present is 4